Move categorize functions into their own file
